유저별, 장치별 통계에 작업시간 추가.

+++ dev/backend/haomibo/src/main/java/com/nuctech/ecuritycheckitem/models/response/userstatistics/HandExaminationStatisticsForPreview.java
@@ -15,7 +15,7 @@ public class HandExaminationStatisticsForPreview {
     String axisLabel;
     long id;
     long time;
-    long workingSeconds;
+    double workingSeconds;
     long totalHandExamination;
     long seizureHandExamination;
     double seizureHandExaminationRate;
